ensure `services` attribute exists immediately
Refactor the way the user record gets written to MongoDB in such a way that the user passed to any callbacks that are invoked (i.e., `afterCreateUser`) contain all of the necessary information to work with other LG services.

This has a side-benefit of one fewer round-trips to the database, also.